Bot/requirements.txt
logging4
aiogram==3.0.0b7
python-dotenv
redis